requirements.txt
azure-functions
pyodbc
python-dateutil
